Major update to version 1.6.0. Older saves are still backwards compatible. Numerous changes and additions: Added about 53 new pokemon, including evolutions of existing species, Regieleki and Regidrago, Meltan and Melmetal, Lucaro, Galvantula, Chandelure, Volcarona, Goodra & Hisuian Goodra and Frosmoth; Added 70+ new abilities; Added additional fields to the BattleMove struct to cut down on the mess of macros (isPunchingMove, damage category, etc.) along with adding and distributing 60+ new moves; Updated/balanced some moves and added a new damage category, variable, which makes the move utilize your highest attacking stat (i.e Hyper Beam is now either a Special move if your Sp. Atk is higher, or a Physical move if your Atk is higher, stat boosts are considered in this calc); Added some new maps to give Regieleki/drago a place to live, updated various other maps; added support for Hidden Abilities and bottle caps; fix Southern Island encounter issue; added 12 new items - 4 evolutions stones, 5 held items, the ability patch, bottle cap, and gold bottle cap, revamped the Battle Frontier to take into account all these new changes, moves, and pokemon, utilize pokemon sets with specific abilities, and support facing Frontier Brains in Doubles battle in the facilities that offer a Doubles option; level_up_learnsets now utilize an actual structure instead of the bitshifting method due to the IDs of the new moves; All new species have pokedex entries, and spare/unused save space was utilized to add more dex entry space to accomodate the new species; new evolution methods were added, including Map Evos (i.e level up Magneton inside New Mauville for Magnezone), seperate Item Male/Female evolutions, and Evo in Rain; distributed all species (new and old) such that all Pokemon in the game can be obtained on way or another; Reworked Battle Dome styles to utilize unused styles and calculate move styles on the fly instead of having a large array; Enigma Berry is its own berry per its Gen 4+ implementation while still supporting e-Reader berries; added a chance for a reward after rebattling Wally in Victory Road; added support for varying fonts for names that are otherwise too long; added a last-used ball menu and use-ball button to the left side of the battle menu, and when selecting a move the ball menu slides away to display a move info menu with its associated button; various other changes, additions, and bug fixes.

+++ include/overworld.h
@@ -73,7 +73,7 @@ extern u8 gLocalLinkPlayerId;
 extern u8 gFieldLinkPlayerCount;
 extern bool8 gSysPcFromPokenav;
 extern u8 gTimeOfDay;
-extern u8 gTimeUpdateCounter;
+extern s16 gTimeUpdateCounter;
 extern struct TimeBlendSettings currentTimeBlend;
 
 // Exported ROM declarations
